Refatoração do código e mascára de input do CPF
