+++ app/signin.js
@@ -1,7 +1,9 @@
 import React, { useState } from 'react';
 import { Stack } from 'expo-router';
-import { View, Text, TouchableOpacity, StyleSheet, Alert } from 'react-native';
-import { TextInput } from 'react-native-paper'
+import { View, Text, TouchableOpacity, Alert } from 'react-native';
+import { styles } from '../styles/authStyles'
+import { validateEmail } from '../utils/validators'
+import { EmailInput, PasswordInput } from '../components/authInputs'
 
 const Signin = () => {
   const [email, setEmail] = useState('');
@@ -49,15 +51,6 @@ const Signin = () => {
     setPassword('');
   };
 
-  const toggleViewPassword = () => {
-    setViewPassword(!viewPassword);
-  }
-
-  const validateEmail = (email) => {
-    const emailRegex = /^\w+@\w+\.\w+$/;
-    return emailRegex.test(email)
-  };
-
   return (
     <View style={styles.container}>
       <Stack.Screen
@@ -72,40 +65,19 @@ const Signin = () => {
       </View>
 
       <View style={styles.inputsContainer}>
-        <TextInput
-          style={styles.input}
-          label="Email"
-          value={email}
-          onChangeText={(textInput) => setEmail(textInput)}
-          inputMode='email'
-          mode='outlined'
-          activeOutlineColor='#189A46'
-          theme={{ roundness: 50 }} 
-          error={errors.email}
+        <EmailInput 
+          email={email}
+          emailError={errors.email}
+          setEmail={setEmail}
         />
 
-        <TextInput
-          style={styles.input}
-          label="Senha"
-          secureTextEntry={!viewPassword}
-          value={password}
-          onChangeText={(textInput) => setPassword(textInput)}
-          mode='outlined'
-          right={ viewPassword ? 
-            <TextInput.Icon icon="eye" 
-              onPress={toggleViewPassword} 
-              style={styles.passwordIcon}
-            /> 
-            : 
-            <TextInput.Icon icon="eye-off" 
-              onPress={toggleViewPassword} 
-              style={styles.passwordIcon}
-            />
-          }
-          activeOutlineColor='#189A46'
-          theme={{ roundness: 50 }} 
-          error={errors.password}
-          onBlur={() => setViewPassword(false)}
+        <PasswordInput
+          label='Senha'
+          password={password}
+          setPassword={setPassword}
+          viewPassword={viewPassword}
+          setViewPassword={setViewPassword}
+          passwordError={errors.password}
         />
       </View>
 
@@ -114,83 +86,11 @@ const Signin = () => {
       </TouchableOpacity>
       
       <TouchableOpacity onPress={() => console.log("esqueci minha senha")}>
-        <Text style={styles.forgotPassword}>Esqueci minha senha</Text>
+        <Text style={styles.link}>Esqueci minha senha</Text>
       </TouchableOpacity>
 
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    height: '100%',
-    width: '100%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 40,
-    paddingBottom: 100
-  },
-  titleContainer: {
-    width: '90%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 2,
-    marginBottom: 40,
-  },
-  inputsContainer: {
-    width: '90%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 8,
-    marginBottom: 24,
-  },
-  title: {
-    fontSize: 32,
-    fontWeight: '600',
-  },
-  subtitle: {
-    fontSize: 20,
-    fontWeight: '500',
-  },
-  input: {
-    width: '100%',
-    height: 50,
-    fontSize: 18,
-    paddingLeft: 10,
-    lineHeight: 24,
-  },
-  passwordIcon: {
-    marginTop: 12,
-  },
-  button: {
-    backgroundColor: 'blue',
-    width: '90%',
-    height: 50,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: 5,
-  },
-  button:{
-    backgroundColor: '#189A46',
-    paddingVertical: 10,
-    borderRadius: 4,
-    alignItems: 'center',
-    borderWidth: 2,
-    borderColor: '#fff',
-    width: '90%',
-    borderRadius: 50,
-  },
-  buttonText:{
-    fontWeight: '500',
-    fontSize: 20,
-    color: '#fff',
-  },
-  forgotPassword: {
-    fontSize: 16,
-    fontWeight: '500',
-    color: '#189A46',
-  }
-});
 
 export default Signin;
